backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pyjwt[crypto]
openai
langsmith
pydantic-settings
python-multipart
